requirements.txt

wheel
gunicorn

Flask
Flask-Caching
Flask-Markdown
#Flask-Appconfig
#Flask-Bootstrap
#Flask-Nav
#Flask-Debug
#flask-wtf


pandas
numpy
pangres>=4.1.1
sqlalchemy
#sqlite3

requests

pyyaml

ratelimit
backoff
